exceptions

+++ src/main/java/com/moviesapplication/moviesApp/service/MovieService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.moviesapplication.moviesApp.exception.blankException;
 import com.moviesapplication.moviesApp.model.MovieModel;
 import com.moviesapplication.moviesApp.repository.MovieRepository;
 
@@ -16,22 +17,49 @@ public class MovieService implements ServiceInt {
 	private MovieRepository movieRepo;
 
 	@Transactional
-	public void addMovie(MovieModel movie) {
-		movieRepo.saveMovie(movie);
-		System.out.println("se agrego una nueva pelicula");
+	public void addMovie(MovieModel movie) throws blankException {
+		
+		if(movie.getName().isBlank() || movie.getDescription().isBlank()) {
+			
+			throw new blankException("Debes ingresar todos los items.....");
+			
+		} else {
+			
+			movieRepo.saveMovie(movie);
+			System.out.println("se agrego una nueva pelicula");
+			
+		}
 		
 	}
 
 	@Transactional
-	public void delMovie(MovieModel movie) {
-		movieRepo.deleteMovie(movie);
-		System.out.println("se elimino la pelicula");
+	public void delMovie(MovieModel movie) throws blankException {
+		if(movie.getName().isBlank()) {
+			
+			throw new blankException("No has ingresado el nombre de la pelicula.....");
+			
+		} else {
+			
+			movieRepo.deleteMovie(movie);
+			System.out.println("se elimino la pelicula");
+			
+		}
+		
 	}
 
 	@Transactional
-	public void edMovie(MovieModel movie) {
-		movieRepo.editMovie(movie);
-		System.out.println("se edito la pelicula");
+	public void edMovie(MovieModel movie) throws blankException {
+		
+		if(movie.getName().isBlank() || movie.getDescription().isBlank()) {
+			
+			throw new blankException("Debes ingresar todos los items.....");
+			
+		} else {
+			
+			movieRepo.editMovie(movie);
+			System.out.println("se edito la pelicula");
+			
+		}
 		
 	}
 
@@ -41,8 +69,17 @@ public class MovieService implements ServiceInt {
 	}
 
 	@Transactional
-	public List<MovieModel> listGenre(String genre) {
-		return movieRepo.moviesPerGenre(genre);
+	public List<MovieModel> listGenre(String genre) throws blankException {
+		if(genre.isBlank()) {
+			
+			throw new blankException("Debes ingresar el genero.....");
+			
+		} else {
+			
+			return movieRepo.moviesPerGenre(genre);
+			
+		}
+		
 	}
 
 }
